LayoutTests/platform/ios/imported/w3c/web-platform-tests/screen-orientation/unlock-expected.txt


Harness Error (FAIL), message = Test named 'unlock() aborts a pending lock request across documents' specified 1 'cleanup' function, and 1 failed.

PASS unlock() doesn't throw when there is no lock
PASS unlock() returns a void value
PASS unlock() doesn't throw when there is no lock with fullscreen
PASS unlock() aborts a pending lock request
FAIL unlock() aborts a pending lock request across documents promise_test: Unhandled rejection with value: object "TypeError: Type error"
